requirements/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/requirements.in -o requirements/requirements.txt
aiohttp==3.9.3
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.1
    # via aiohttp
alembic==1.13.1
    # via optuna
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
attrs==23.2.0
    # via
    #   aiohttp
    #   flake8-annotations
    #   flake8-bugbear
backoff==2.2.1
    # via dspy-ai
bandit==1.7.8
    # via flake8-bandit
bcrypt==4.1.2
    # via passlib
black==24.3.0
    # via flake8-black
boltons==23.1.1
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   uvicorn
colorlog==6.8.2
    # via optuna
coverage==7.4.4
    # via pytest-cov
cryptography==42.0.5
    # via python-jose
darglint==1.8.1
datasets==2.14.7
    # via dspy-ai
defusedxml==0.7.1
dill==0.3.7
    # via
    #   datasets
    #   multiprocess
distlib==0.3.8
    # via virtualenv
distro==1.9.0
    # via openai
dspy-ai==2.4.0
ecdsa==0.18.0
    # via python-jose
fastapi==0.110.0
filelock==3.13.1
    # via
    #   huggingface-hub
    #   virtualenv
flake8==7.0.0
    # via
    #   flake8-annotations
    #   flake8-bandit
    #   flake8-black
    #   flake8-bugbear
    #   flake8-docstrings
flake8-annotations==3.0.1
flake8-bandit==4.1.1
flake8-black==0.3.6
flake8-bugbear==24.2.6
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.10.0
    # via
    #   datasets
    #   huggingface-hub
greenlet==3.0.3
    # via sqlalchemy
gunicorn==21.2.0
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.4
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via openai
huggingface-hub==0.21.4
    # via datasets
identify==2.5.35
    # via pre-commit
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
joblib==1.3.2
    # via dspy-ai
mako==1.3.2
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via mako
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.15
    # via datasets
mypy==1.9.0
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nodeenv==1.8.0
    # via pre-commit
numpy==1.26.4
    # via
    #   datasets
    #   optuna
    #   pandas
    #   pyarrow
openai==1.14.2
    # via dspy-ai
optuna==3.6.0
    # via dspy-ai
packaging==24.0
    # via
    #   black
    #   datasets
    #   gunicorn
    #   huggingface-hub
    #   optuna
    #   pytest
pandas==2.2.1
    # via
    #   datasets
    #   dspy-ai
passlib==1.7.4
pathspec==0.12.1
    # via black
pbr==6.0.0
    # via stevedore
platformdirs==4.2.0
    # via
    #   black
    #   virtualenv
pluggy==1.4.0
    # via pytest
pre-commit==3.6.2
pyarrow==15.0.2
    # via datasets
pyarrow-hotfix==0.6
    # via datasets
pyasn1==0.5.1
    # via
    #   python-jose
    #   rsa
pycodestyle==2.11.1
    # via
    #   flake8
    #   flake8-import-order
pycparser==2.21
    # via cffi
pydantic==2.5.0
    # via
    #   dspy-ai
    #   fastapi
    #   openai
pydantic-core==2.14.1
    # via pydantic
pydocstyle==6.3.0
    # via flake8-docstrings
pyflakes==3.2.0
    # via flake8
pygments==2.17.2
    # via rich
pytest==8.1.1
    # via pytest-cov
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via uvicorn
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   bandit
    #   datasets
    #   huggingface-hub
    #   optuna
    #   pre-commit
    #   uvicorn
regex==2023.12.25
    # via dspy-ai
requests==2.31.0
    # via
    #   datasets
    #   dspy-ai
    #   fsspec
    #   huggingface-hub
rich==13.7.1
    # via bandit
rsa==4.9
    # via python-jose
ruff==0.3.4
setuptools==69.2.0
    # via
    #   flake8-import-order
    #   nodeenv
shellcheck-py==0.10.0.1
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
snowballstemmer==2.2.0
    # via pydocstyle
sqlalchemy==2.0.29
    # via
    #   alembic
    #   optuna
starlette==0.36.3
    # via fastapi
stevedore==5.2.0
    # via bandit
toml==0.10.2
tqdm==4.66.2
    # via
    #   datasets
    #   dspy-ai
    #   huggingface-hub
    #   openai
    #   optuna
types-toml==0.10.8.20240310
typing-extensions==4.10.0
    # via
    #   alembic
    #   fastapi
    #   huggingface-hub
    #   mypy
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2024.1
    # via pandas
ujson==5.9.0
    # via dspy-ai
urllib3==2.2.1
    # via requests
uvicorn==0.29.0
uvloop==0.19.0
    # via uvicorn
virtualenv==20.25.1
    # via pre-commit
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
xxhash==3.4.1
    # via datasets
yarl==1.9.4
    # via aiohttp
